+++ wooting-macro-backend/Cargo.toml
@@ -39,8 +39,5 @@ rayon = "1.10"
 url = "2.5"
 uuid = { version = "1.8", features = ["fast-rng", "v4", "macro-diagnostics"] }
 
-[target.'cfg(target_os = "windows")'.dependencies]
-multiinput = { git = "https://github.com/simon-wh/multiinput-rust", branch = "master" }
-
 [profile.release]
 lto = true
