+++ wooting-macro-backend/src/grabbing/mod.rs
@@ -1,4 +1,3 @@
 pub mod executor;
-pub mod linux;
 pub mod macos;
-pub mod windows;
+pub mod windows_linux;
