--- wooting-macro-backend/src/grabbing/windows/mod.rs
@@ -1,2 +1,0 @@
-pub mod input;
-mod matcher;
