+++ wooting-macro-backend/src/grabbing/windows_linux/input.rs
@@ -1,22 +1,17 @@
-#[cfg(target_os = "windows")]
+#[cfg(any(target_os = "windows", target_os = "linux"))]
 pub mod input {
     use std::sync::atomic::{AtomicBool, Ordering};
     use std::sync::Arc;
     use std::time;
 
     use itertools::Itertools;
-    use log::*;
-    use multiinput::*;
     use rdev::EventType;
     use tokio::sync::mpsc::UnboundedSender;
 
     use crate::grabbing::executor::input::MacroExecutorEvent;
-    use crate::grabbing::windows::matcher::input::{
-        check_macro_execution_simply, check_macro_execution_simply_async,
-    };
+    use crate::grabbing::windows_linux::matcher::input::check_macro_execution_simply;
     use crate::hid_table::*;
     use crate::macros::macro_data::MacroLookup;
-    use crate::plugin::delay::DEFAULT_DELAY;
 
     use crate::RwLock;
 
@@ -26,14 +21,6 @@ pub mod input {
         map: Arc<RwLock<MacroLookup>>,
     ) {
         tokio::time::sleep(time::Duration::from_millis(3000)).await;
-
-        let mut manager = RawInputManager::new().unwrap();
-        manager.register_devices(DeviceType::Keyboards);
-        warn!("Device list {:#?}", manager.get_device_list());
-
-        for device in manager.get_device_list().keyboards.iter() {
-            warn!("Device status {:#?}", device);
-        }
 
         let og_previously_pressed_keys: Arc<RwLock<Vec<u32>>> = Arc::new(RwLock::from(vec![]));
         let og_current_pressed_keys: Arc<RwLock<Vec<u32>>> = Arc::new(RwLock::from(vec![]));
@@ -95,6 +82,7 @@ pub mod input {
                                 &schan_macro_execute_inner,
                             );
 
+                            // We always send the release event of a key, since there's no harm to it.
                             Some(event)
                         }
                         EventType::ButtonPress(button) => {
@@ -122,60 +110,5 @@ pub mod input {
                 }
             })
         });
-    }
-
-    pub async fn check_keypress_simon_multiinput(
-        inner_is_listening: Arc<AtomicBool>,
-        schan_macro_execute: UnboundedSender<MacroExecutorEvent>,
-        map: Arc<RwLock<MacroLookup>>,
-    ) {
-        tokio::time::sleep(time::Duration::from_millis(3000)).await;
-
-        let mut manager = RawInputManager::new().unwrap();
-        manager.register_devices(DeviceType::Keyboards);
-        warn!("Device list {:#?}", manager.get_device_list());
-
-        for device in manager.get_device_list().keyboards.iter() {
-            warn!("Device status {:#?}", device);
-        }
-
-        let mut previously_pressed_keys: Vec<u32> = vec![];
-        let mut current_pressed_keys: Vec<u32> = vec![];
-
-        loop {
-            if inner_is_listening.load(Ordering::Relaxed) {
-                if let Some(event) = manager.get_event() {
-                    if let RawEvent::KeyboardEvent(_, key, event) = event {
-                        match event {
-                            State::Pressed => {
-                                current_pressed_keys
-                                    .push(*MULTIINPUT_TO_HID.get(&key).unwrap_or(&0));
-                                current_pressed_keys =
-                                    current_pressed_keys.into_iter().unique().collect();
-
-                                check_macro_execution_simply_async(
-                                    &current_pressed_keys,
-                                    &previously_pressed_keys,
-                                    map.clone(),
-                                    &schan_macro_execute,
-                                )
-                                .await;
-                            }
-                            State::Released => {
-                                previously_pressed_keys.clone_from(&current_pressed_keys);
-
-                                current_pressed_keys
-                                    .retain(|x| x != MULTIINPUT_TO_HID.get(&key).unwrap_or(&0));
-                            }
-                        }
-                    }
-                }
-                tokio::time::sleep(time::Duration::from_millis(DEFAULT_DELAY)).await;
-                // thread::sleep(time::Duration::from_millis(DEFAULT_DELAY));
-            } else {
-                // thread::sleep(time::Duration::from_millis(2000));
-                tokio::time::sleep(time::Duration::from_millis(2000)).await;
-            }
-        }
     }
 }
