+++ wooting-macro-backend/src/grabbing/windows_linux/matcher.rs
@@ -1,0 +1,1 @@
+#[cfg(any(target_os = "windows", target_os = "linux"))]
