+++ wooting-macro-backend/src/hid_table.rs
@@ -4,8 +4,6 @@ use std::hash::Hash;
 use crate::plugin::mouse::MouseButton;
 use lazy_static::lazy_static;
 
-#[cfg(target_os = "windows")]
-use multiinput::KeyId;
 use rdev;
 use rdev::Key;
 
@@ -417,360 +415,6 @@ pub static ref RDEV_TO_HID: HashMap<Key, u32> = {
         scancode.insert(Key::ShiftRight, 0xe5); //SHIFT_RIGHT
         scancode.insert(Key::Alt, 0xe6); //ALT_RIGHT
         scancode.insert(Key::MetaRight, 0xe7); //META_RIGHT
-
-    scancode
-};
-
-}
-
-#[cfg(target_os = "windows")]
-lazy_static! {
-#[derive(Debug, PartialEq, Hash, std::cmp::Eq)]
-pub static ref MULTIINPUT_TO_HID: HashMap<KeyId, u32> = {
-    let mut scancode: HashMap<KeyId, u32> = HashMap::new();
-        scancode.insert(KeyId::A, 0x04); //US_A
-        scancode.insert(KeyId::B, 0x05); //US_B
-        scancode.insert(KeyId::C, 0x06); //US_C
-        scancode.insert(KeyId::D, 0x07); //US_D
-        scancode.insert(KeyId::E, 0x08); //US_E
-        scancode.insert(KeyId::F, 0x09); //US_F
-        scancode.insert(KeyId::G, 0x0a); //US_G
-        scancode.insert(KeyId::H, 0x0b); //US_H
-        scancode.insert(KeyId::I, 0x0c); //US_I
-        scancode.insert(KeyId::J, 0x0d); //US_J
-        scancode.insert(KeyId::K, 0x0e); //US_K
-        scancode.insert(KeyId::L, 0x0f); //US_L
-        scancode.insert(KeyId::M, 0x10); //US_M
-        scancode.insert(KeyId::N, 0x11); //US_N
-        scancode.insert(KeyId::O, 0x12); //US_O
-        scancode.insert(KeyId::P, 0x13); //US_P
-        scancode.insert(KeyId::Q, 0x14); //US_Q
-        scancode.insert(KeyId::R, 0x15); //US_R
-        scancode.insert(KeyId::S, 0x16); //US_S
-        scancode.insert(KeyId::T, 0x17); //US_T
-        scancode.insert(KeyId::U, 0x18); //US_U
-        scancode.insert(KeyId::V, 0x19); //US_V
-        scancode.insert(KeyId::W, 0x1a); //US_W
-        scancode.insert(KeyId::X, 0x1b); //US_X
-        scancode.insert(KeyId::Y, 0x1c); //US_Y
-        scancode.insert(KeyId::Z, 0x1d); //US_Z
-
-        scancode.insert(KeyId::N1, 0x1e); //DIGIT1
-        scancode.insert(KeyId::N2, 0x1f); //DIGIT2
-        scancode.insert(KeyId::N3, 0x20); //DIGIT3
-        scancode.insert(KeyId::N4, 0x21); //DIGIT4
-        scancode.insert(KeyId::N5, 0x22); //DIGIT5
-        scancode.insert(KeyId::N6, 0x23); //DIGIT6
-        scancode.insert(KeyId::N7, 0x24); //DIGIT7
-        scancode.insert(KeyId::N8, 0x25); //DIGIT8
-        scancode.insert(KeyId::N9, 0x26); //DIGIT9
-        scancode.insert(KeyId::N0, 0x27); //DIGIT0
-
-        scancode.insert(KeyId::Return, 0x28); //ENTER
-        scancode.insert(KeyId::Escape, 0x29); //ESCAPE
-        scancode.insert(KeyId::Back, 0x2a); //BACKSPACE
-        scancode.insert(KeyId::Tab, 0x2b); //TAB
-        scancode.insert(KeyId::Space, 0x2c); //SPACE
-        scancode.insert(KeyId::OEMMinus, 0x2d); //MINUS
-        scancode.insert(KeyId::OEMPlus, 0x2e); //EQUAL
-        scancode.insert(KeyId::OEM4, 0x2f); //BRACKET_LEFT
-
-        scancode.insert(KeyId::OEM6, 0x30); //BRACKET_RIGHT
-        scancode.insert(KeyId::OEM5, 0x31); //BACKSLASH
-
-        //scancode.insert(0x32, 0x0000); //INTL_HASH
-
-        scancode.insert(KeyId::OEM1, 0x33); //SEMICOLON
-        scancode.insert(KeyId::OEM7, 0x34); //QUOTE
-        scancode.insert(KeyId::OEM3, 0x35); //BACKQUOTE
-        scancode.insert(KeyId::OEMComma, 0x36); //COMMA
-        scancode.insert(KeyId::OEMPeriod, 0x37); //PERIOD
-        scancode.insert(KeyId::OEM2, 0x38); //SLASH
-        scancode.insert(KeyId::CapsLock, 0x39); //CAPS_LOCK
-
-        scancode.insert(KeyId::F1, 0x3a); //F1
-        scancode.insert(KeyId::F2, 0x3b); //F2
-        scancode.insert(KeyId::F3, 0x3c); //F3
-        scancode.insert(KeyId::F4, 0x3d); //F4
-        scancode.insert(KeyId::F5, 0x3e); //F5
-        scancode.insert(KeyId::F6, 0x3f); //F6
-        scancode.insert(KeyId::F7, 0x40); //F7
-        scancode.insert(KeyId::F8, 0x41); //F8
-        scancode.insert(KeyId::F9, 0x42); //F9
-        scancode.insert(KeyId::F10, 0x43); //F10
-        scancode.insert(KeyId::F11, 0x44); //F11
-        scancode.insert(KeyId::F12, 0x45); //F12
-        scancode.insert(KeyId::F13, 0x68); //F13
-        scancode.insert(KeyId::F14, 0x69); //F14
-        scancode.insert(KeyId::F15, 0x6a); //F15
-        scancode.insert(KeyId::F16, 0x6b); //F16
-        scancode.insert(KeyId::F17, 0x6c); //F17
-        scancode.insert(KeyId::F18, 0x6d); //F18
-        scancode.insert(KeyId::F19, 0x6e); //F19
-        scancode.insert(KeyId::F20, 0x6f); //F20
-        scancode.insert(KeyId::F21, 0x70); //F21
-        scancode.insert(KeyId::F22, 0x71); //F22
-        scancode.insert(KeyId::F23, 0x72); //F23
-        scancode.insert(KeyId::F24, 0x73); //F24
-
-        scancode.insert(KeyId::VolumeMute, 0x7f); //VOLUME_MUTE
-        scancode.insert(KeyId::VolumeDown, 0x81); //VOLUME_DOWN
-        scancode.insert(KeyId::VolumeUp, 0x80); //VOLUME_UP
-
-        scancode.insert(KeyId::Print, 0x46); //PRINT_SCREEN
-        scancode.insert(KeyId::ScrollLock, 0x47); //SCROLL_LOCK
-
-        scancode.insert(KeyId::Pause, 0x48); //PAUSE
-        scancode.insert(KeyId::Insert, 0x49); //INSERT
-        scancode.insert(KeyId::Home, 0x4a); //HOME
-        scancode.insert(KeyId::Prior, 0x4b); //PAGE_UP
-        scancode.insert(KeyId::Delete, 0x4c); //DEL
-        scancode.insert(KeyId::End, 0x4d); //END
-        scancode.insert(KeyId::Next, 0x4e); //PAGE_DOWN
-
-        scancode.insert(KeyId::Right, 0x4f); //ARROW_RIGHT
-        scancode.insert(KeyId::Left, 0x50); //ARROW_LEFT
-        scancode.insert(KeyId::Down, 0x51); //ARROW_DOWN
-        scancode.insert(KeyId::Up, 0x52); //ARROW_UP
-
-        scancode.insert(KeyId::NumLock, 0x53); //NUM_LOCK
-        scancode.insert(KeyId::NumpadDivide, 0x54); //NUMPAD_DIVIDE
-        scancode.insert(KeyId::NumpadMultiply, 0x55); //NUMPAD_MULTIPLY
-        scancode.insert(KeyId::NumpadSubtract, 0x56); //NUMPAD_SUBTRACT
-        scancode.insert(KeyId::NumpadAdd, 0x57); //NUMPAD_ADD
-        scancode.insert(KeyId::NumpadEnter, 0x58); //NUMPAD_ENTER
-
-        scancode.insert(KeyId::Numpad1, 0x59); //NUMPAD1
-        scancode.insert(KeyId::Numpad2, 0x5a); //NUMPAD2
-        scancode.insert(KeyId::Numpad3, 0x5b); //NUMPAD3
-        scancode.insert(KeyId::Numpad4, 0x5c); //NUMPAD4
-        scancode.insert(KeyId::Numpad5, 0x5d); //NUMPAD5
-        scancode.insert(KeyId::Numpad6, 0x5e); //NUMPAD6
-        scancode.insert(KeyId::Numpad7, 0x5f); //NUMPAD7
-        scancode.insert(KeyId::Numpad8, 0x60); //NUMPAD8
-        scancode.insert(KeyId::Numpad9, 0x61); //NUMPAD9
-        scancode.insert(KeyId::Numpad0, 0x62); //NUMPAD0
-        scancode.insert(KeyId::NumpadDecimal, 0x85); //NUMPAD_COMMA (KP_DELETE)
-
-        scancode.insert(KeyId::OEM5, 0x64); //INTL_BACKSLASH
-        //TODO: scancode.insert(KeyId::Function, 0x65); //CONTEXT_MENU
-
-        // scancode.insert(0x67, Key::Key); //NUMPAD_EQUAL
-
-        // scancode.insert(0x74, Key::Key); //OPEN
-        //
-        // scancode.insert(0x75, Key::Key); //HELP
-        //
-        // scancode.insert(0x66, Key::Key); //POWER
-        // scancode.insert(0x77, 0x0000); //SELECT
-        //
-        // scancode.insert(0x79, Key::Key); //AGAIN
-        // scancode.insert(0x7a, Key::Key); //UNDO
-        // scancode.insert(0x7b, Key::Key); //CUT
-        // scancode.insert(0x7c, Key::Key); //COPY
-        // scancode.insert(0x7d, Key::Key); //PASTE
-        // scancode.insert(0x7e, Key::Key); //FIND
-        // scancode.insert(0x7f, Key::Key); //VOLUME_MUTE
-        //
-        // scancode.insert(0x80, Key::Key); //VOLUME_UP
-        // scancode.insert(0x81, Key::Key); //VOLUME_DOWN
-        // scancode.insert(0x85, Key::Key); //NUMPAD_COMMA
-        //
-        // scancode.insert(0x87, Key::Key); //INTL_RO
-        // scancode.insert(0x88, Key::Key); //KANA_MODE
-        // scancode.insert(0x89, Key::Key); //INTL_YEN
-        // scancode.insert(0x8a, Key::Key); //CONVERT
-        // scancode.insert(0x8b, Key::Key); //NON_CONVERT
-        // scancode.insert(0x90, Key::Key); //LANG1
-        // scancode.insert(0x91, Key::Key); //LANG2
-        // scancode.insert(0x92, Key::Key); //LANG3
-        // scancode.insert(0x93, Key::Key); //LANG4
-
-        scancode.insert(KeyId::LeftControl, 0xe0); //CONTROL_LEFT
-        scancode.insert(KeyId::LeftShift, 0xe1); //SHIFT_LEFT
-        scancode.insert(KeyId::LeftMenu, 0xe2); //ALT_LEFT
-        scancode.insert(KeyId::LeftWindows, 0xe3); //META_LEFT
-        scancode.insert(KeyId::RightControl, 0xe4); //CONTROL_RIGHT
-        scancode.insert(KeyId::RightShift, 0xe5); //SHIFT_RIGHT
-        scancode.insert(KeyId::RightMenu, 0xe6); //ALT_RIGHT
-        scancode.insert(KeyId::RightWindows, 0xe7); //META_RIGHT
-
-    scancode
-};
-
-}
-
-#[cfg(target_os = "windows")]
-lazy_static! {
-///Conversion from HID codes to the library backend enums.
-#[derive(Debug, PartialEq, Hash, std::cmp::Eq)]
-pub static ref HID_TO_MULTIINPUT: HashMap<u32, KeyId> = {
-    let mut scancode: HashMap<u32, KeyId> = HashMap::new();
-        scancode.insert(0x04, KeyId::A); //US_A
-        scancode.insert(0x05, KeyId::B); //US_B
-        scancode.insert(0x06, KeyId::C); //US_C
-        scancode.insert(0x07, KeyId::D); //US_D
-        scancode.insert(0x08, KeyId::E); //US_E
-        scancode.insert(0x09, KeyId::F); //US_F
-        scancode.insert(0x0a, KeyId::G); //US_G
-        scancode.insert(0x0b, KeyId::H); //US_H
-        scancode.insert(0x0c, KeyId::I); //US_I
-        scancode.insert(0x0d, KeyId::J); //US_J
-        scancode.insert(0x0e, KeyId::K); //US_K
-        scancode.insert(0x0f, KeyId::L); //US_L
-        scancode.insert(0x10, KeyId::M); //US_M
-        scancode.insert(0x11, KeyId::N); //US_N
-        scancode.insert(0x12, KeyId::O); //US_O
-        scancode.insert(0x13, KeyId::P); //US_P
-        scancode.insert(0x14, KeyId::Q); //US_Q
-        scancode.insert(0x15, KeyId::R); //US_R
-        scancode.insert(0x16, KeyId::S); //US_S
-        scancode.insert(0x17, KeyId::T); //US_T
-        scancode.insert(0x18, KeyId::U); //US_U
-        scancode.insert(0x19, KeyId::V); //US_V
-        scancode.insert(0x1a, KeyId::W); //US_W
-        scancode.insert(0x1b, KeyId::X); //US_X
-        scancode.insert(0x1c, KeyId::Y); //US_Y
-        scancode.insert(0x1d, KeyId::Z); //US_Z
-
-        scancode.insert(0x1e, KeyId::N1); //DIGIT1
-        scancode.insert(0x1f, KeyId::N2); //DIGIT2
-        scancode.insert(0x20, KeyId::N3); //DIGIT3
-        scancode.insert(0x21, KeyId::N4); //DIGIT4
-        scancode.insert(0x22, KeyId::N5); //DIGIT5
-        scancode.insert(0x23, KeyId::N6); //DIGIT6
-        scancode.insert(0x24, KeyId::N7); //DIGIT7
-        scancode.insert(0x25, KeyId::N8); //DIGIT8
-        scancode.insert(0x26, KeyId::N9); //DIGIT9
-        scancode.insert(0x27, KeyId::N0); //DIGIT0
-
-        scancode.insert(0x28, KeyId::Return); //ENTER
-        scancode.insert(0x29, KeyId::Escape); //ESCAPE
-        scancode.insert(0x2a, KeyId::Back); //BACKSPACE
-        scancode.insert(0x2b, KeyId::Tab); //TAB
-        scancode.insert(0x2c, KeyId::Space); //SPACE
-        scancode.insert(0x2d, KeyId::OEMMinus); //MINUS
-        scancode.insert(0x2e, KeyId::OEMPlus); //EQUAL
-        scancode.insert(0x2f, KeyId::OEM4); //BRACKET_LEFT
-
-        scancode.insert(0x30, KeyId::OEM6); //BRACKET_RIGHT
-        scancode.insert(0x31, KeyId::OEM5); //BACKSLASH
-
-        //scancode.insert(0x32, 0x0000); //INTL_HASH
-
-        scancode.insert(0x33, KeyId::OEM1); //SEMICOLON
-        scancode.insert(0x34, KeyId::OEM7); //QUOTE
-        scancode.insert(0x35, KeyId::OEM3); //BACKQUOTE
-        scancode.insert(0x36, KeyId::OEMComma); //COMMA
-        scancode.insert(0x37, KeyId::OEMPeriod); //PERIOD
-        scancode.insert(0x38, KeyId::OEM2); //SLASH
-        scancode.insert(0x39, KeyId::CapsLock); //CAPS_LOCK
-
-        scancode.insert(0x3a, KeyId::F1); //F1
-        scancode.insert(0x3b, KeyId::F2); //F2
-        scancode.insert(0x3c, KeyId::F3); //F3
-        scancode.insert(0x3d, KeyId::F4); //F4
-        scancode.insert(0x3e, KeyId::F5); //F5
-        scancode.insert(0x3f, KeyId::F6); //F6
-
-        scancode.insert(0x40, KeyId::F7); //F7
-        scancode.insert(0x41, KeyId::F8); //F8
-        scancode.insert(0x42, KeyId::F9); //F9
-        scancode.insert(0x43, KeyId::F10); //F10
-        scancode.insert(0x44, KeyId::F11); //F11
-        scancode.insert(0x45, KeyId::F12); //F12
-        scancode.insert(0x68, KeyId::F13); //F13
-        scancode.insert(0x69, KeyId::F14); //F14
-        scancode.insert(0x6a, KeyId::F15); //F15
-        scancode.insert(0x6b, KeyId::F16); //F16
-        scancode.insert(0x6c, KeyId::F17); //F17
-        scancode.insert(0x6d, KeyId::F18); //F18
-        scancode.insert(0x6e, KeyId::F19); //F19
-        scancode.insert(0x6f, KeyId::F20); //F20
-        scancode.insert(0x70, KeyId::F21); //F21
-        scancode.insert(0x71, KeyId::F22); //F22
-        scancode.insert(0x72, KeyId::F23); //F23
-        scancode.insert(0x73, KeyId::F24); //F24
-
-        scancode.insert(0x7f, KeyId::VolumeMute); //VOLUME_MUTE
-        scancode.insert(0x81, KeyId::VolumeDown); //VOLUME_DOWN
-        scancode.insert(0x80, KeyId::VolumeUp); //VOLUME_UP
-
-        scancode.insert(0x46, KeyId::Print); //PRINT_SCREEN
-        scancode.insert(0x47, KeyId::ScrollLock); //SCROLL_LOCK
-
-        scancode.insert(0x48, KeyId::Pause); //PAUSE
-        scancode.insert(0x49, KeyId::Insert); //INSERT
-        scancode.insert(0x4a, KeyId::Home); //HOME
-        scancode.insert(0x4b, KeyId::Prior); //PAGE_UP
-        scancode.insert(0x4c, KeyId::Delete); //DEL
-        scancode.insert(0x4d, KeyId::End); //END
-        scancode.insert(0x4e, KeyId::Next); //PAGE_DOWN
-
-        scancode.insert(0x4f, KeyId::Right); //ARROW_RIGHT
-        scancode.insert(0x50, KeyId::Left); //ARROW_LEFT
-        scancode.insert(0x51, KeyId::Down); //ARROW_DOWN
-        scancode.insert(0x52, KeyId::Up); //ARROW_UP
-
-        scancode.insert(0x53, KeyId::NumLock); //NUM_LOCK
-        scancode.insert(0x54, KeyId::NumpadDivide); //NUMPAD_DIVIDE
-        scancode.insert(0x55, KeyId::NumpadMultiply); //NUMPAD_MULTIPLY
-        scancode.insert(0x56, KeyId::NumpadSubtract); //NUMPAD_SUBTRACT
-        scancode.insert(0x57, KeyId::NumpadAdd); //NUMPAD_ADD
-        scancode.insert(0x58, KeyId::NumpadEnter); //NUMPAD_ENTER
-
-        scancode.insert(0x59, KeyId::Numpad1); //NUMPAD1
-        scancode.insert(0x5a, KeyId::Numpad2); //NUMPAD2
-        scancode.insert(0x5b, KeyId::Numpad3); //NUMPAD3
-        scancode.insert(0x5c, KeyId::Numpad4); //NUMPAD4
-        scancode.insert(0x5d, KeyId::Numpad5); //NUMPAD5
-        scancode.insert(0x5e, KeyId::Numpad6); //NUMPAD6
-        scancode.insert(0x5f, KeyId::Numpad7); //NUMPAD7
-        scancode.insert(0x60, KeyId::Numpad8); //NUMPAD8
-        scancode.insert(0x61, KeyId::Numpad9); //NUMPAD9
-        scancode.insert(0x62, KeyId::Numpad0); //NUMPAD0
-        scancode.insert(0x85, KeyId::NumpadDecimal); //NUMPAD_COMMA (KP_DELETE)
-
-        scancode.insert(0x64, KeyId::OEM5); //INTL_BACKSLASH
-        //TODO: scancode.insert(0x65, KeyId::Function); //CONTEXT_MENU
-
-        // scancode.insert(0x67, Key::Key); //NUMPAD_EQUAL
-        // scancode.insert(0x74, Key::Key); //OPEN
-        //
-        // scancode.insert(0x75, Key::Key); //HELP
-        //
-        // scancode.insert(0x66, Key::Key); //POWER
-        // scancode.insert(0x77, 0x0000); //SELECT
-        //
-        // scancode.insert(0x79, Key::Key); //AGAIN
-        // scancode.insert(0x7a, Key::Key); //UNDO
-        // scancode.insert(0x7b, Key::Key); //CUT
-        // scancode.insert(0x7c, Key::Key); //COPY
-        // scancode.insert(0x7d, Key::Key); //PASTE
-        // scancode.insert(0x7e, Key::Key); //FIND
-
-
-
-        // scancode.insert(0x85, Key::Key); //NUMPAD_COMMA (KP_DELETE)
-        //
-        // scancode.insert(0x87, Key::Key); //INTL_RO
-        // scancode.insert(0x88, Key::Key); //KANA_MODE
-        // scancode.insert(0x89, Key::Key); //INTL_YEN
-        // scancode.insert(0x8a, Key::Key); //CONVERT
-        // scancode.insert(0x8b, Key::Key); //NON_CONVERT
-        // scancode.insert(0x90, Key::Key); //LANG1
-        // scancode.insert(0x91, Key::Key); //LANG2
-        // scancode.insert(0x92, Key::Key); //LANG3
-        // scancode.insert(0x93, Key::Key); //LANG4
-
-        scancode.insert(0xe0, KeyId::LeftControl); //CONTROL_LEFT
-        scancode.insert(0xe1, KeyId::LeftShift); //SHIFT_LEFT
-        scancode.insert(0xe2, KeyId::LeftMenu); //ALT_LEFT
-        scancode.insert(0xe3, KeyId::LeftWindows); //META_LEFT
-        scancode.insert(0xe4, KeyId::RightControl); //CONTROL_RIGHT
-        scancode.insert(0xe5, KeyId::RightShift); //SHIFT_RIGHT
-        scancode.insert(0xe6, KeyId::RightMenu); //ALT_RIGHT
-        scancode.insert(0xe7, KeyId::RightWindows); //META_RIGHT
 
     scancode
 };
